organize tests better
